tx : fix pricePoint check failed

+++ src/rpc/core/rpcclient.cpp
@@ -282,7 +282,7 @@ Array RPCConvertValues(const string &strMethod, const vector<string> &strParams)
 
     /* for cdp */
     if (strMethod == "submitpricefeedtx"        && n > 1) ConvertTo<Array>(params[1]);
-    if (strMethod == "submitpricefeedtx"        && n > 2) ConvertTo<double>(params[2]);
+    if (strMethod == "submitpricefeedtx"        && n > 2) ConvertTo<uint64_t>(params[2]);
 
     if (strMethod == "submitstakefcointx" && n > 3) ConvertTo<uint64_t>(params[3]);
     if (strMethod == "submitstakefcointx" && n > 4) ConvertTo<uint64_t>(params[4]);
